Moved drawing to lib
