+++ SignumGenerator/SignumLib/Drawing/SignumBaseDrawFiguresHonorary.cs
@@ -1,8 +1,9 @@
 ﻿using System.Drawing;
 using System.Drawing.Drawing2D;
-using SignumGenerator.Helpers;
+using SignumLib.Base;
+using SignumLib.Helpers;
 
-namespace SignumGenerator.Signum
+namespace SignumLib.Drawing
 {
     /*
      * Drawing Honorary heraldic figures
