+++ SignumGenerator/SignumLib/Helpers/DdsImageParser.cs
@@ -1,14 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing.Imaging;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
 
-namespace SignumGenerator.Helpers
+namespace SignumLib.Helpers
 {
     #region DDSImage Class
     public class DDSImage : IDisposable
